Dev -> master

+++ interactions/buttons/stop/stopseaheaven.js
@@ -5,20 +5,29 @@ module.exports = {
 	async execute(interaction) {
 		const { NodeSSH } = require("node-ssh");
 		const ssh = new NodeSSH();
-		ssh
-			.connect({
-				host: "dev.codesj.kr",
-				username: "codesj",
-				password: sshpw,
-				readyTimeout: 30000,
-			})
-			.then(function () {
-				ssh
-					.execCommand(`echo ${sshpw} | sudo -S systemctl stop seaheaven`, {})
-					.then(function (result) {
-						ssh.execCommand(sshpw, {});
-					});
+		if (
+			interaction.member.roles.cache.some((r) => r.id == "1074241811314389022")
+		) {
+			ssh
+				.connect({
+					host: "dev.codesj.kr",
+					username: "codesj",
+					password: sshpw,
+					readyTimeout: 30000,
+				})
+				.then(function () {
+					ssh
+						.execCommand(`echo ${sshpw} | sudo -S systemctl stop seaheaven`, {})
+						.then(function (result) {
+							ssh.execCommand(sshpw, {});
+						});
+				});
+		} else {
+			await interaction.reply({
+				content: "권한이 없습니다. 관리자에게 요청하세요",
+				ephemeral: true,
 			});
+		}
 
 		await interaction.reply({
 			content: "SeaheavenWiki 중지 완료",
